% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/root.R
\name{root}
\alias{root}
\title{Compute the nth root. Computes the square root by default.}
\usage{
root(x, n = 2)
}
\arguments{
\item{x}{The vector to be rooted.}

\item{n}{the degree of the root. Default value is 2.}
}
\value{
A vector that is the nth root of \code{x}.
}
\description{
Compute the nth root. Computes the square root by default.
}
\details{
Using the internal \code{\link{pow}} function, the nth root is computed as
\code{x^(1/n)}.
}
\examples{
root(9)
root(8, 3)
}
